requirements.txt
annotated-types==0.5.0
ansi2html==1.8.0
anyio==3.7.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dnspython==2.3.0
fastapi==0.100.0
Flask==2.2.5
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
nest-asyncio==1.5.6
numpy==1.25.1
packaging==23.1
pandas==2.0.3
plotly==5.15.0
pydantic==2.0.2
pydantic_core==2.1.2
pymongo==4.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
retrying==1.3.4
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tenacity==8.2.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
Werkzeug==2.2.3